src/backend/requirements.txt
flask
requests
django
djangorestframework
django-mongodb-backend
pymongo
serializers